# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sweet Cakes - Premium Bakery API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from "./product";

export interface CartItem {
  productId: string;
  product: Product;
  quantity: number;
  subtotal: number;
}