# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sweet Cakes - Premium Bakery API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderRequestPaymentMethod } from "./createOrderRequestPaymentMethod";
import type { ShippingAddress } from "./shippingAddress";

export interface CreateOrderRequest {
  shippingAddress: ShippingAddress;
  paymentMethod: CreateOrderRequestPaymentMethod;
}