# lib/api-zod/src/generated/types/shippingAddress.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sweet Cakes - Premium Bakery API
 * OpenAPI spec version: 0.1.0
 */

export interface ShippingAddress {
  fullName: string;
  phone: string;
  address: string;
  city: string;
  state: string;
  zipCode: string;
}